requirements.txt
boto3==1.23.10
botocore==1.26.10
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.34
pandas==1.4.2
numpy == 1.22.4
PyYAML==6.0
scikit-learn == 1.1.1
Flask==2.1.1
pymysql==1.0.2
pytest == 7.1.2